+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     ProjectQ.
 -   Support for parameterized quantum gates (e.g. Rx, Ry, Rz, QubitOperator, TimeEvolution, etc.)
 -   Support for exporting HiQ-ProjectQ quantum circuits as JSON encoded strings
+-   Support for reading/writing quantum circuits using OpenQASM
 
 ### Updated